# internal/docs/openapi.yaml
openapi: 3.1.0
info:
  title: socialstats API
  version: 1.0.0
  summary: Public view counts for a social video URL.
  description: |
    Give the service a video URL and it returns that video's public metrics.

    YouTube, TikTok and Meta (Facebook + Instagram) are implemented.

    **Reliability differs by platform.** YouTube is read through the official
    Data API and is reliable. Neither TikTok nor Meta offers an API for
    arbitrary video URLs, so their counters come from the public pages.

    - TikTok serves an anti-bot challenge to a substantial share of requests,
      which the service retries. Expect occasional `502 upstream_blocked` on
      TikTok URLs, and retry — it usually succeeds.
    - Instagram is read from its login-free embed render and is reasonably
      reliable. Facebook prefers the Graph API when a token is configured and
      can see the object, and otherwise reads the public page — for which
      Facebook often serves a login wall instead. That is also
      `502 upstream_blocked`, but retrying it will not help.

    **There is a dashboard too.** The same binary serves a server-rendered UI at
    `/`, sharing these endpoints' service layer, session cookies and CSRF token.
    This document covers the JSON API only; the HTML pages are not part of the
    contract and may change freely.

    **Two ways to authenticate, and the choice matters.** A session cookie or an
...
    second. Cookie-authenticated writes must echo the CSRF token; bearer and
...
servers:
...
paths:
  /v1/stats:
    get:
      tags: [stats]
      summary: Get stats for a video URL
      description: |
        Resolves the URL to a platform, fetches the video's public metrics and
        returns them in a platform-agnostic shape.
      operationId: getStats
      parameters:
        - name: url
          in: query
          required: true
          description: |
            The video URL. Any common form is accepted, with or without a scheme.

            YouTube: `watch?v=`, `youtu.be/`, `/shorts/`, `/embed/`, `/live/`,
            `/v/`, `youtube-nocookie.com`.

            TikTok: `/@user/video/<id>`, `/@user/photo/<id>`, `/v/<id>.html`,
            and the `vm.tiktok.com` / `vt.tiktok.com` / `/t/<token>` short links
            (resolved by following redirects).

            Instagram: `/p/<code>`, `/reel/<code>`, `/reels/<code>`,
            `/tv/<code>`, with or without a leading handle.

            Facebook: `/watch/?v=<id>`, `/reel/<id>`, `/<page>/videos/<id>`,
            `/<page>/videos/<slug>/<id>`, `video.php?v=<id>`, and the
            `fb.watch/<token>` / `/share/v/<token>` short links.
          schema:
            type: string
            format: uri
            examples:
              - https://www.youtube.com/watch?v=dQw4w9WgXcQ
              - https://youtu.be/dQw4w9WgXcQ
              - https://www.youtube.com/shorts/dQw4w9WgXcQ
              - https://www.tiktok.com/@thebillionairebros/video/7249376077976472833
              - https://www.instagram.com/reel/Cx1y2z3AbCd/
              - https://www.facebook.com/watch/?v=1234567890
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200': { $ref: '#/components/responses/Stats' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
        '502': { $ref: '#/components/responses/UpstreamError' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }

    post:
      tags: [stats]
      summary: Get stats for a video URL (JSON body)
      description: |
        Identical to `GET /v1/stats`, for URLs that are awkward to pass in a
        query string. Unknown body fields are rejected.
      operationId: postStats
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatsRequest' }
            examples:
              youtube:
                summary: A YouTube watch URL
                value: { url: 'https://www.youtube.com/watch?v=dQw4w9WgXcQ' }
              short:
                summary: A YouTube Shorts URL
                value: { url: 'https://www.youtube.com/shorts/dQw4w9WgXcQ' }
      responses:
        '200': { $ref: '#/components/responses/Stats' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
        '502': { $ref: '#/components/responses/UpstreamError' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
  /v1/videos:
    get:
      tags: [tracking]
      summary: The caller's tracked videos
      description: |
        Returns the caller's list with each video's current counters and its
        growth over the requested window.

        `views_gained` is the current view count minus the last reading at or
        before the window's start. It is **absent rather than zero** when there
        is no baseline — a video added yesterday has no week-old reading — and
        it is **signed**, because platforms revise view counts downward and a
        negative is a real measurement.
      operationId: listVideos
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: window
          in: query
          description: How far back the growth baseline is taken from. Max one year.
          schema: { type: string, default: 168h, examples: ['24h', '168h'] }
        - name: platform
          in: query
          schema: { $ref: '#/components/schemas/Platform' }
        - name: sort
          in: query
          schema:
            type: string
            enum: [views, gained, recent, title, fetched]
            default: views
        - name: sparkline
          in: query
          description: |
            Attach this many recent readings to each entry, oldest first. Off by
            default: it costs one query per video, which is fine for a page and
            wasteful for a client that only wants the numbers.
          schema: { type: integer, minimum: 0, maximum: 200, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: The caller's tracked videos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoListEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      tags: [tracking]
      summary: Track a video
      description: |
        Resolves the URL, records the video and starts tracking it for the
        caller — then **fetches it immediately**, so the response carries real
        numbers rather than an empty row that fills in within the next interval.

        A failed first fetch still tracks the video and still returns `201`. The
        failure is recorded on the row (`schedule.last_fetch_status`) and in the
        fetch log, and the poller retries: a TikTok challenge at 14:03 is not a
        reason to refuse to track a video.

        The exception is short links (`vm.tiktok.com`, `fb.watch`,
        `/share/...`), which carry no id until the redirect is followed. A
        failed fetch there means the video genuinely cannot be identified, and
        the request fails rather than tracking something unknown.
      operationId: trackVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrackRequest' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/TrackRequest' }
      responses:
        '201':
          description: |
            Now tracked. Whether the first fetch succeeded is reported by the
            entry's own fetch status, not by the status code.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/LimitReached' }
        '502': { $ref: '#/components/responses/UpstreamError' }
  /v1/videos/{id}:
    parameters:
      - $ref: '#/components/parameters/VideoID'
    get:
      tags: [tracking]
      summary: One tracked video
      operationId: getVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: The tracked video.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }

    patch:
      tags: [tracking]
      summary: Rename or annotate a tracked video
      description: |
        Changes only the fields that are sent. Omitting `notes` leaves the
        existing notes alone; sending `""` clears them.

        Both are per-user: two accounts tracking the same video keep their own
        label and notes.
      operationId: updateTrackedVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTrackingRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }

    delete:
      tags: [tracking]
      summary: Stop tracking a video
      description: |
        Removes the video from the caller's list. The video and its entire
        history survive: other accounts may be tracking it, and re-adding it
        later restores both the collected history and the label.

        Polling stops only when the last tracker leaves.
      operationId: untrackVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: Untracked.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      untracked: { type: boolean }
                      history_kept: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }
  /v1/videos/{id}/history:
    get:
      tags: [tracking]
      summary: A video's readings over time
      description: |
        Returns the time series for one video.

        `source` says where the figures came from, because the two are not
        equivalent: `snapshots` are individual measurements, `daily_rollup` rows
        summarise days whose raw data has since been expired. A `day` bucket
        reads the rollup when one exists and falls back to the raw series when
        it does not, rather than returning an empty chart that looks like the
        video has no history.

        Bucketing takes the **last** reading in each bucket, not an average:
        these are cumulative counters, and the mean of two of them is not a
        number the video ever had.
      operationId: getVideoHistory
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/VideoID'
        - name: from
          in: query
          description: RFC 3339. Defaults to seven days before `to`.
          schema: { type: string, format: date-time }
        - name: to
          in: query
          description: RFC 3339. Defaults to now.
          schema: { type: string, format: date-time }
        - name: bucket
          in: query
          schema: { type: string, enum: [raw, hour, day], default: raw }
      responses:
        '200':
          description: The video's readings.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HistoryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }
  /v1/videos/{id}/attempts:
    get:
      tags: [tracking]
      summary: A video's recent fetch attempts
      description: |
        Every attempt to refresh this video, successful or not. This is how
        "why is this number stale?" gets answered without reading logs.
      operationId: getVideoAttempts
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/VideoID'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        '200':
          description: Recent attempts, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttemptsEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }
  /v1/videos/{id}/refresh:
    post:
      tags: [tracking]
      summary: Bring the next fetch forward
      description: |
        Schedules the video to be fetched on the next poll tick, for a caller
        who does not want to wait for the interval.

        It returns `202` because it queues rather than fetches. Fetching inline
        on demand would let one impatient account spend the whole deployment's
        TikTok and Meta budget, and the poller is already what paces platform
        access.
      operationId: refreshVideo
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/VideoID'
      responses:
        '202':
          description: Queued for the next poll tick.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      queued: { type: boolean }
                      video: { $ref: '#/components/schemas/Entry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotTracked' }
        '410':
          description: |
            The video has been retired: the platform said it no longer exists,
            repeatedly. Its history is still readable; it is simply not fetched
            any more.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /v1/dashboard/summary:
    get:
      tags: [tracking]
      summary: Headline figures for the caller
      description: |
        Totals, a per-platform breakdown, coverage, and the videos that moved
        most over the window.

        `stale` counts videos whose newest reading is older than twice their own
        fetch interval — per video rather than one cutoff, because a six-hourly
        YouTube video and a twelve-hourly TikTok video go stale at different
        speeds and a single threshold would misreport one of them.
      operationId: getDashboardSummary
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: window
          in: query
          schema: { type: string, default: 168h }
        - name: movers
          in: query
          schema: { type: integer, minimum: 0, maximum: 50, default: 5 }
      responses:
        '200':
          description: The caller's headline figures.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SummaryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Create an account
      description: |
        Creates an account and signs it in. Accepts JSON or a form post.

        Passwords must be at least 12 characters and are rejected if they appear
        in the embedded breach list or contain the address's local part. There
        are no composition rules: they produce `Password1!` and reject
        `correct horse battery staple`, which is the wrong way round.

        A duplicate address returns `409`, which does disclose that the address
        already has an account. Avoiding that requires accepting the sign-up
        silently and mailing the address instead, and this service sends no
        mail — so the disclosure is documented rather than pretended away.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { $ref: '#/components/responses/SignedIn' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Sign in
      description: |
        Verifies a password and issues a session. Accepts JSON or a form post.

        Every failure returns the same `401 invalid_credentials`, whether the
        address is unknown, the password is wrong or the account is suspended —
        and each path costs the same time, because an unknown address is still
        hashed against a decoy. A faster "no such user" would be an
        account-existence oracle that identical wording does not hide.

        Failed attempts are limited per address and per source address, five per
        fifteen minutes each. A `429` carries `Retry-After` computed from the
        bucket's real refill rate. A successful sign-in clears the counter.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
          application/x-www-form-urlencoded:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/SignedIn' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: End this session
      description: |
        Revokes the session the request was made with and clears the cookies.

        Succeeds whether or not there was a session. Signing out is an
        intention, and "you were not signed in" gives the caller nothing to act
        on.
      operationId: logout
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: Signed out.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedOutEnvelope' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/auth/logout-all:
    post:
      tags: [auth]
      summary: End every session for this account
      description: |
        Revokes every session the account has, on every device.

        This is what server-side sessions buy that a stateless signed token
        would not: revocation is a DELETE, not a denylist that has to be
        consulted on every request until the token would have expired anyway.
      operationId: logoutEverywhere
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: Every session was revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedOutEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/auth/password:
    post:
      tags: [auth]
      summary: Change password
      description: |
        Verifies the current password and replaces it.

        Every session is revoked, including the one making the request. The
        usual reason to change a password is believing someone else has it, and
        leaving their session signed in would make the change ceremonial.
      operationId: changePassword
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: Password changed and every session revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedOutEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/auth/me:
    get:
      tags: [auth]
      summary: The signed-in account
      operationId: getMe
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        '200':
          description: The current account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ss_session
      description: |
        The session cookie set by register and login. HttpOnly, SameSite=Lax and
        Secure outside development.

        Requests authenticated this way must also pass the CSRF check on unsafe
        methods: echo the readable CSRF cookie back in `X-CSRF-Token` or a
        `csrf_token` form field.
    bearerToken:
      type: http
      scheme: bearer
      description: |
        The same session token, sent as `Authorization: Bearer <token>`. Login
        returns it in the body for exactly this purpose.

        Bearer requests are exempt from CSRF, and not as a convenience: CSRF
        exists because browsers attach cookies to cross-site requests
        automatically. They do not attach this header on a third party's behalf,
        so the request cannot be forged.

        A request with an `application/json` body is exempt for the same kind of
        reason: an HTML form can only send urlencoded, multipart or text/plain,
        so a JSON content type is itself proof the request did not come from
        one. Every form-encoded write is checked, including sign-in and sign-up.
  parameters:
...
    RequestID:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Correlation id. When supplied it is echoed back and used in logs;
        otherwise the service generates one.
      schema:
        type: string
...
  schemas:
    Platform:
      type: string
      description: A supported social video platform.
      enum: [youtube, meta, tiktok]
...
  responses:
    Stats:
      description: The video's metrics.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
        X-Cache: { $ref: '#/components/headers/XCache' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/StatsEnvelope' }
          examples:
            youtube:
              summary: A YouTube video
              value:
                data:
                  platform: youtube
                  video_id: dQw4w9WgXcQ
                  canonical_url: https://www.youtube.com/watch?v=dQw4w9WgXcQ
                  title: Never Gonna Give You Up
                  channel_id: UCuAXFkgsw1L7xaCfnd5JJOw
                  channel_title: Rick Astley
                  published_at: '2009-10-25T06:57:33Z'
                  view_count: 1500000000
                  like_count: 18000000
                  comment_count: 2200000
                  fetched_at: '2026-08-27T10:00:00Z'
                  cached: false
            tiktok:
              summary: A TikTok video, which also reports shares and saves
              value:
                data:
                  platform: tiktok
                  video_id: '7249376077976472833'
                  canonical_url: https://www.tiktok.com/@thebillionairebros/video/7249376077976472833
                  title: Jim Simmons, the founder of Renaissance Technologies...
                  channel_id: '7205953225726624770'
                  channel_title: TheBillionaireBros
                  published_at: '2023-06-27T14:41:42Z'
                  view_count: 220500
                  like_count: 8702
                  comment_count: 113
                  share_count: 690
                  save_count: 2732
                  fetched_at: '2026-08-27T11:29:20Z'
                  cached: false
            instagram:
              summary: An Instagram reel
              value:
                data:
                  platform: meta
                  video_id: Cx1y2z3AbCd
                  canonical_url: https://www.instagram.com/reel/Cx1y2z3AbCd/
                  title: Behind the scenes on assignment in Patagonia
                  channel_id: '787132'
                  channel_title: natgeo
                  published_at: '2023-06-27T14:41:42Z'
                  view_count: 220500
                  like_count: 8702
                  comment_count: 113
                  fetched_at: '2026-08-27T11:31:04Z'
                  cached: false
            facebook:
              summary: A public Facebook video
              value:
                data:
                  platform: meta
                  video_id: '1234567890'
                  canonical_url: https://www.facebook.com/watch/?v=1234567890
                  title: A public video
                  channel_id: '555'
                  channel_title: National Geographic
                  published_at: '2023-06-27T14:41:42Z'
                  view_count: 98765
                  like_count: 4321
                  comment_count: 42
                  share_count: 17
                  fetched_at: '2026-08-27T11:33:47Z'
                  cached: false
            hiddenCounters:
              summary: Uploader hides likes and comments — the fields are absent
              value:
                data:
                  platform: youtube
                  video_id: dQw4w9WgXcQ
                  canonical_url: https://www.youtube.com/watch?v=dQw4w9WgXcQ
                  title: A quiet upload
                  view_count: 1024
                  fetched_at: '2026-08-27T10:00:00Z'
                  cached: true
...
    BadRequest:
      description: The request was malformed, or the URL is not a supported video URL.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missingURL:
              summary: No url supplied
              value:
                error:
                  code: missing_url
                  message: query parameter 'url' is required
                  request_id: 09f67ff774aefae24185e16d344b930b
            invalidURL:
              summary: Not a parseable video URL
              value:
                error:
                  code: invalid_url
                  message: the supplied url could not be parsed as a supported video url
                  request_id: 09f67ff774aefae24185e16d344b930b
            unsupportedPlatform:
              summary: A host no provider claims
              value:
                error:
                  code: unsupported_platform
                  message: no provider handles that url
                  request_id: add2d0a65dc565fa18585bb50d243bc3
    NotFound:
      description: The video, or the endpoint, does not exist.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            video:
              value:
                error:
                  code: not_found
                  message: the video does not exist, is private, or was removed
                  request_id: 82386569467a5dad423f6d1f90278a05
...
    InternalError:
      description: An unexpected server fault. Quote the request id when reporting it.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...